requirements.txt
absl-py
addict
aiohttp
aiosignal
ansi2html
async-timeout
attrs
blinker
cachetools
cattrs
certifi
charset-normalizer
chumpy
click
ConfigArgParse
contourpy
coremltools
cvxopt
cycler
dadaptation
daqp
dash
dash-core-components
dash-html-components
dash-table
docker-pycreds
dotmap
ecos
exceptiongroup
fastjsonschema
filelock
Flask
fonttools
frozenlist
fsspec
gitdb
GitPython
google-auth
google-auth-oauthlib
grpcio
h5py
idna
imageio
importlib-metadata
importlib-resources
itsdangerous
Jinja2
joblib
jsonschema
jsonschema-specifications
jupyter_core
keyboard
kiwisolver
lightning
lightning-utilities
Markdown
MarkupSafe
matplotlib
mpmath
multidict
nbformat
nest-asyncio
networkx
numpy
oauthlib
open3d
opencv-python
osqp
packaging
pandas
pathtools
Pillow
platformdirs
plotly
prodigyopt
promise
protobuf
psutil
pyaml
pyasn1
pyasn1-modules
pyDeprecate
pygame
PyOpenGL
pyparsing
PyQt5
PyQt5_sip
pyqtgraph
pyquaternion
pyserial
python-dateutil
pytorch-lightning
pytz
PyYAML
qdldl
qpsolvers
quadprog
referencing
requests
requests-oauthlib
retrying
rpds-py
rsa
scikit-learn
scipy
scs
seaborn
sentry-sdk
setproctitle
shortuuid
six
smmap
sympy
tenacity
tensorboard
tensorboard-data-server
termcolor
threadpoolctl
torch
torchmetrics
tqdm
traitlets
transforms3d
typing_extensions
tzdata
urllib3
vctoolkit
wandb
Werkzeug
yarl
yaspin
zipp